// types/singlePageSEOTypes.ts
export interface singlePageSEOProps {
  singlePageTitle: string;
  singlePageExcerpt: string;
  singlePageSectionName: string;
  singlePageSectionSlug: string;
  singlePageCategoryName: string;
  singlePageCategorySlug: string;
  singlePageSlug: string;
  singlePageDate: string;
  singlePageModifiedDate: string;
  singlePageFeaturedImage: string;
  singlePageAuthor: string;
}